First pass of vme_fake bridge driver
This is the first pass commit of the vme_fake bridge driver. It
currently doesn't support DMA, but master & slave windows, the
location monitor and "VME Interrupts" are supported/emulated.

Due to vme_bridge.h (understandably) not being packaged in the kernel
headers provided by Ubuntu, it is included to make this build
out-of-tree.

Signed-off-by: Martyn Welch <martyn@welchs.me.uk>